*   Applied Andreas' tcpnal compiler optimization bugfix patch       from HEAD portals (different way of constructing tcp HELLO       header to avoid pointer aliasing) to lnet
  *   Applied qswnal build fix to lustre-portals.m4 from HEAD
      portals to lnet

  *   lnet version of gmnal running @ HP

  *   fixed bad 64bit cast in acceptor.c

  *   fixed lconf to work with newconfig modules under lnet

+++ lnet/lnet/acceptor.c
@@ -310,7 +310,7 @@ ptl_acceptor(void *arg)
 	__u32          peer_ip;
 	int            peer_port;
         ptl_ni_t      *blind_ni;
-        int            secure = (int)arg;
+        int            secure = (int)((unsigned long)arg);
 
 	LASSERT (ptl_acceptor_state.pta_sock == NULL);
